+++ packages/engine-multi/src/api/lifecycle.ts
@@ -120,13 +120,22 @@ export const log = (
 ) => {
   const { threadId } = event;
 
-  if (event.message.name !== 'JOB') {
-    context.logger.proxy(event.message);
+  if (event.log.name !== 'JOB') {
+    // Forward the log event to the engine's logger
+    // Note that we may have to parse the serialized log string
+    const proxy = {
+      ...event.log,
+      message:
+        typeof event.log.message == 'string'
+          ? JSON.parse(event.log.message)
+          : event.log.message,
+    };
+    context.logger.proxy(proxy);
   }
 
   context.emit(externalEvents.WORKFLOW_LOG, {
     threadId,
-    ...event.message,
+    ...event.log,
   });
 };