+++ packages/engine-multi/src/worker/events.ts
@@ -67,8 +67,14 @@ export interface JobErrorEvent extends InternalEvent {
   next: string[];
 }
 
+export type SerializedLogEvent = Omit<JSONLog, 'message'> & {
+  // the message is either an array of strings/object to log,
+  // or a JSON array that was previously serialized
+  message: string | any[];
+};
+
 export interface LogEvent extends InternalEvent {
-  message: JSONLog;
+  log: SerializedLogEvent;
 }
 
 export interface ErrorEvent {