+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @openfn/cli
 
+## 0.4.16
+
+### Patch Changes
+
+- Updated dependencies [649ca43]
+  - @openfn/logger@0.0.20
+  - @openfn/compiler@0.0.39
+  - @openfn/deploy@0.4.1
+  - @openfn/runtime@0.2.6
+
 ## 0.4.15
 
 ### Patch Changes